Fix struct names
